Consolidate Python modules into ccai.app
